Add files via upload
I implemented the changes that Marissa made with the ArrayLists. This version is not working properly, but I was wondering if maybe one of you could possible figure out how to get the option from the Interest_Activity to be loaded into the Main_Activity. That way when the button is pressed it will ask the user for their interests and then generate the webpage.